separate graph generation and path finding and fix bug in path finding
