+++ .gitignore
@@ -26,6 +26,7 @@ Makefile.in
 /src/frontend/diff-sizes
 /src/frontend/frameify
 /src/frontend/frameplayer
+/src/frontend/graphgen
 /src/frontend/vp8play
 /src/frontend/vp8decode
 /src/frontend/collisions
