+++ src/frontend/Makefile.am
@@ -1,6 +1,6 @@
 AM_CPPFLAGS = -I$(srcdir)/../util -I$(srcdir)/../decoder -I$(srcdir)/../display -I$(srcdir)/../encoder $(X11_XCB_CFLAGS) $(GLU_CFLAGS) $(CXX11_FLAGS)
 AM_CXXFLAGS = $(PICKY_CXXFLAGS) $(NODEBUG_CXXFLAGS)
-bin_PROGRAMS = vp8decode vp8play diff-sizes frameify frameplayer collisions
+bin_PROGRAMS = vp8decode vp8play diff-sizes frameify graphgen frameplayer collisions
 
 vp8decode_SOURCES = vp8decode.cc
 vp8decode_LDADD = ../decoder/libalfalfadecoder.a ../encoder/libalfalfaencoder.a ../display/libalfalfadisplay.a ../util/libalfalfautil.a $(X11_XCB_LIBS) $(GLU_LIBS)
@@ -13,6 +13,9 @@ diff_sizes_LDADD = ../decoder/libalfalfadecoder.a ../encoder/libalfalfaencoder.a
 
 frameify_SOURCES = frameify.cc
 frameify_LDADD = ../decoder/libalfalfadecoder.a ../encoder/libalfalfaencoder.a ../util/libalfalfautil.a
+
+graphgen_SOURCES = graphgen.cc
+graphgen_LDADD = ../decoder/libalfalfadecoder.a ../util/libalfalfautil.a
 
 frameplayer_SOURCES = frameplayer.cc
 frameplayer_LDADD = ../decoder/libalfalfadecoder.a ../encoder/libalfalfaencoder.a ../display/libalfalfadisplay.a ../util/libalfalfautil.a $(X11_XCB_LIBS) $(GLU_LIBS)
